requirements/base.txt
Django==5.1.5
django-crispy-forms
crispy-bootstrap5
django-allauth
django-extensions
django-environ

